refactor(MyReservationCancel): 취소하기 기능 및 컴포넌트 리팩토링
