+++ src/components/pages/myReservation/MyReservationList.tsx
@@ -1,8 +1,14 @@
 import MyReservationCard from '@/src/components/pages/myReservation/MyReservationCard';
 import Button from '@/src/components/primitives/Button';
 import EmptyList from '@/src/components/primitives/EmptyList';
+import ConfirmModal from '@/src/components/primitives/modal/ConfirmModal';
+import { cancelMyReservation } from '@/src/services/pages/myReservation/api';
+import { queries } from '@/src/services/primitives/queries';
 import { MyReservationListResponse } from '@/src/types/myReservationType';
+import { useQueryClient } from '@tanstack/react-query';
+import { AxiosError } from 'axios';
 import { useRouter } from 'next/navigation';
+import { useCallback, useState } from 'react';
 
 interface Props {
   pagesData: MyReservationListResponse[];
@@ -17,8 +23,36 @@ export default function MyReservationList({
   isFetchingNextPage,
   fetchNextPage,
 }: Props) {
+  const queryClient = useQueryClient();
   const router = useRouter();
+  const [modalOpen, setModalOpen] = useState(false);
+  const [reservationId, setReservationId] = useState<number | null>(null);
   const isEmpty = pagesData[0].totalCount === 0;
+
+  const handleOpenCancelModal = (reservationId: number) => {
+    setModalOpen(true);
+    setReservationId(reservationId);
+  };
+
+  const handleCancleReservation = useCallback(async () => {
+    if (reservationId === null) return;
+
+    try {
+      await cancelMyReservation(reservationId);
+      queryClient.invalidateQueries({
+        queryKey: queries.myReservationList(),
+      }); // myReservationList 쿼리키 갖고 있는 캐싱 데이터 전부 초기화
+      setModalOpen(false); // 팝업 닫기
+    } catch (error) {
+      const err = error as AxiosError;
+
+      console.log(err);
+
+      if (err.response?.status === 403) {
+        alert('취소 권한이 없습니다.');
+      }
+    }
+  }, [reservationId]);
 
   return (
     <div>
@@ -33,18 +67,29 @@ export default function MyReservationList({
           </Button>
         </EmptyList>
       ) : (
-        <ul>
-          {pagesData.map((page) =>
-            page.reservations.map((reservation) => (
-              <li
-                key={reservation.id}
-                className='pt-5 pb-[30px] border-t border-t-gray-50 lg:mb-6 lg:pb-0 lg:pt-0 lg:border-b-0'
-              >
-                <MyReservationCard reservation={reservation} />
-              </li>
-            ))
-          )}
-        </ul>
+        <>
+          <ul>
+            {pagesData.map((page) =>
+              page.reservations.map((reservation) => (
+                <li
+                  key={reservation.id}
+                  className='pt-5 pb-[30px] border-t border-t-gray-50 lg:mb-6 lg:pb-0 lg:pt-0 lg:border-b-0'
+                >
+                  <MyReservationCard
+                    reservation={reservation}
+                    onCancel={handleOpenCancelModal}
+                  />
+                </li>
+              ))
+            )}
+          </ul>
+          <ConfirmModal
+            isOpen={modalOpen}
+            message={`예약을 취소하시겠어요?`}
+            onConfirm={handleCancleReservation}
+            onCancel={() => setModalOpen(false)}
+          />
+        </>
       )}
     </div>
   );
